refactor: split core into _internal and core
