+++ _internal/index.ts
@@ -1,0 +1,1 @@
+export { createSubscribeHOC } from './subscribe-hoc'
