+++ _internal/tsconfig.json
@@ -1,0 +1,10 @@
+{
+	"extends": "../tsconfig.json",
+	"compilerOptions": {
+	  "rootDir": "..",
+	  "outDir": "./dist"
+	},
+	"include": ["./**/*.ts"]
+
+}
+
